orders/src/routes/show.ts
import express, { Request, Response } from 'express';
import { requireAuth, NotFoundError, NotAuthorizedError } from '@chm-tickets/common';
import { Order } from '../models/order';

export const showRouter = express.Router();

showRouter.get('/api/orders/:id', requireAuth, async (req: Request, res: Response) => {
  const order = await Order.findById(req.params.id).populate('ticket');

  if (!order) {
    throw new NotFoundError();
  }

  if (order.userId !== req.currentUser!.id) {
    throw new NotAuthorizedError();
  }

  res.send(order);
});
